refactor(convert): extract list_item module with try_convert entry
Phase 1 of fulgur-odgo. Move the three list-item dispatcher branches
from convert_node_inner into convert/list_item.rs, plus the helper
build_list_item_body.

- pub(super) fn try_convert(doc, node_id, ctx, depth) -> Option<...>:
  combines all three branches (Outside marker / display:list-item
  fallback / Inside non-inline-root) preserving exact precedence and
  fall-through semantics.
- fn build_list_item_body (private): moved unchanged.

list_item.rs declares use super::{list_marker, pseudo, positioned} so
the existing qualified calls inside the moved branches resolve as
sibling-submodule paths.

The dispatcher in convert_node_inner now starts with let
width/height computation followed by a single list_item::try_convert
early-return.

No behavioral change. VRT byte-identical against committed goldens.
This is the highest-risk move in the plan; bisect against e128c39 if
regressions appear later.
